trying to track down rounding error in phase

+++ crone/ugens/CutFadeVoice/interp.h
@@ -5,13 +5,13 @@
 #ifndef CUTFADELOOP_INTERP_H
 #define CUTFADELOOP_INTERP_H
 
-inline float cubicinterp(float x, float y0, float y1, float y2, float y3)
+inline double cubicinterp(double x, double y0, double y1, double y2, double y3)
 {
-    // 4-point, 3rd-order Hermite (x-form)
-    float c0 = y1;
-    float c1 = 0.5f * (y2 - y0);
-    float c2 = y0 - 2.5f * y1 + 2.f * y2 - 0.5f * y3;
-    float c3 = 0.5f * (y3 - y0) + 1.5f * (y1 - y2);
+    // 4-point, 3rd-order Hermite (x-orm)
+    double c0 = y1;
+    double c1 = 0.5 * (y2 - y0);
+    double c2 = y0 - 2.5 * y1 + 2. * y2 - 0.5 * y3;
+    double c3 = 0.5 * (y3 - y0) + 1.5 * (y1 - y2);
 
     return ((c3 * x + c2) * x + c1) * x + c0;
 }
